requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-openai==0.3.11
langchain-postgres==0.0.13
langchain-text-splitters==0.3.7
langsmith==0.3.19
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.69.0
orjson==3.10.16
packaging==24.2
pgvector==0.3.6
propcache==0.3.1
psycopg==3.2.6
psycopg-pool==3.2.6
psycopg2==2.9.10
pycryptodome==3.15.0
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
pypdf==5.4.0
python-dotenv==1.1.0
python-multipart==0.0.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.14.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
yarl==1.18.3
zstandard==0.23.0
